Move class creation to separate file. Add list of original property names to keep track of names once annotations @PrimaryKey and @ColumnInfo are read
