+++ androidutilities-room/src/main/java/com/heyzeusv/androidutilities/room/RoomProcessor.kt
@@ -1,7 +1,6 @@
 package com.heyzeusv.androidutilities.room
 
 import com.google.devtools.ksp.processing.Dependencies
-import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
@@ -9,11 +8,6 @@ import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.validate
 import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.FunSpec
-import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.PropertySpec
-import com.squareup.kotlinpoet.TypeSpec
-import com.squareup.kotlinpoet.ksp.toTypeName
 
 class RoomProcessor(private val environment: SymbolProcessorEnvironment) : SymbolProcessor {
     private val logger = environment.logger
@@ -57,62 +51,6 @@ class RoomProcessor(private val environment: SymbolProcessorEnvironment) : Symbo
         val ret = symbols.filterNot { it.validate() }.toList()
         return ret
     }
-}
-
-private fun recreateClass(
-    classDeclaration: KSClassDeclaration,
-    logger: KSPLogger,
-): TypeSpec.Builder {
-    val classBuilder = TypeSpec
-        .classBuilder(classDeclaration.utilName())
-        .addModifiers(KModifier.DATA)
-    val constructorBuilder = FunSpec.constructorBuilder()
-
-    classDeclaration.getAllProperties().forEach { prop ->
-        prop.qualifiedName?.getShortName()?.let { name ->
-            val annotationNames = prop.annotations.map { it.shortName.getShortName() }
-            if (annotationNames.contains("Ignore")) {
-                logger.info("Ignoring $name of type ${prop.type}")
-            } else if (annotationNames.contains("Embedded")) {
-                logger.info("Embedded class ${prop.type}")
-                val embeddedClass = prop.type.resolve().declaration as KSClassDeclaration
-                classBuilder.handleEmbeddedClass(constructorBuilder, embeddedClass, logger)
-            } else {
-                constructorBuilder.addParameter(name, prop.type.toTypeName())
-                classBuilder.addProperty(
-                    PropertySpec.builder(name, prop.type.toTypeName()).initializer(name).build()
-                )
-            }
-        }
-    }
-
-    return classBuilder.primaryConstructor(constructorBuilder.build())
-}
-
-// TODO: Handle ColumnInfo annotation
-private fun TypeSpec.Builder.handleEmbeddedClass(
-    constructorBuilder: FunSpec.Builder,
-    classDeclaration: KSClassDeclaration,
-    logger: KSPLogger,
-): TypeSpec.Builder {
-    classDeclaration.getAllProperties().forEach { prop ->
-        prop.qualifiedName?.getShortName()?.let { name ->
-            val annotationNames = prop.annotations.map { it.shortName.getShortName() }
-            if (annotationNames.contains("Ignore")) {
-                logger.info("Ignoring $name of type ${prop.type}")
-            } else if (annotationNames.contains("Embedded")) {
-                logger.info("Embedded class ${prop.type}")
-                val embeddedClass = prop.type.resolve().declaration as KSClassDeclaration
-                this.handleEmbeddedClass(constructorBuilder, embeddedClass, logger)
-            } else {
-                constructorBuilder.addParameter(name, prop.type.toTypeName())
-                this.addProperty(
-                    PropertySpec.builder(name, prop.type.toTypeName()).initializer(name).build()
-                )
-            }
-        }
-    }
-    return this
 }
 
 fun KSClassDeclaration.utilName(): String = "${simpleName.getShortName()}RoomUtil"
